Move loosely related code from image.cc
Dirive-by: cleanup img_ops.h
