+++ lib/jxl_lists.cmake
@@ -213,6 +213,7 @@ set(JPEGXL_INTERNAL_DEC_SOURCES
   jxl/image_bundle.h
   jxl/image_metadata.cc
   jxl/image_metadata.h
+  jxl/image_ops.cc
   jxl/image_ops.h
   jxl/inverse_mtf-inl.h
   jxl/lehmer_code.h
